move deformable examples to a single folder and rename them; change license to google 2016
