+++ examples/CMakeLists.txt
@@ -1,4 +1,4 @@
-SUBDIRS( HelloWorld BasicDemo )
+SUBDIRS( HelloWorld BasicDemo)
 IF(BUILD_BULLET3)
 	SUBDIRS( ExampleBrowser RobotSimulator SharedMemory ThirdPartyLibs/Gwen ThirdPartyLibs/BussIK ThirdPartyLibs/clsocket OpenGLWindow TwoJoint )
 ENDIF()
