+++ examples/ExampleBrowser/premake4.lua
@@ -183,9 +183,6 @@ project "App_BulletExampleBrowser"
 		"../VoronoiFracture/*",
 		"../SoftDemo/*",
 		"../DeformableDemo/*",
-		"../DeformableContact/*",
-		"../VolumetricDeformable/*",
-		"../Pinch/*",
 		"../RollingFrictionDemo/*",
 		"../rbdl/*",
 		"../FractureDemo/*",
